Move nupic/math/* to nupic/utils/
