+++ src/CMakeLists.txt
@@ -121,13 +121,6 @@ set(engine_files
     nupic/engine/Watcher.cpp
     nupic/engine/Watcher.hpp
 )
-  
-set(math_files
-    nupic/math/StlIo.cpp
-    nupic/math/StlIo.hpp
-    nupic/math/Topology.cpp
-    nupic/math/Topology.hpp
-)
 
 
 set(ntypes_files
@@ -202,6 +195,10 @@ set(utils_files
     nupic/utils/SdrMetrics.hpp
     nupic/utils/MurmurHash3.hpp
     nupic/utils/MurmurHash3.cpp
+    nupic/utils/StlIo.cpp
+    nupic/utils/StlIo.hpp
+    nupic/utils/Topology.cpp
+    nupic/utils/Topology.hpp
 )
 
 set(examples_files
@@ -215,7 +212,6 @@ set(examples_files
 source_group("algorithms" FILES ${algorithm_files})
 source_group("encoders" FILES ${encoders_files})
 source_group("engine" FILES ${engine_files})
-source_group("math" FILES ${math_files})
 source_group("ntypes" FILES ${ntypes_files})
 source_group("os" FILES ${os_files})
 source_group("regions" FILES ${regions_files})
@@ -238,7 +234,6 @@ add_library(${src_objlib} OBJECT
     ${algorithm_files} 
     ${encoders_files} 
     ${engine_files} 
-    ${math_files} 
     ${ntypes_files} 
     ${os_files} 
     ${regions_files}
