+++ src/nupic/utils/Topology.hpp
